Executable OpenAI/requirements.txt
Flask
openai==0.28
requests
instabot
Pillow
